Separate the login screen in a separated file and add validation to display or biometric design.
